+++ Shopipy.Persistence/Models/ProductVariation.cs
@@ -1,27 +1,26 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
 
-namespace Shopipy.Persistence.Models
+namespace Shopipy.Persistence.Models;
+
+public class ProductVariation
 {
-    public class ProductVariation
-    {
-        [Key]
-        public int VariationId { get; set; }
+    [Key]
+    public int VariationId { get; set; }
 
-        [ForeignKey("Product")]
-        public required int ProductId { get; set; }
+    [ForeignKey("Product")]
+    public required int ProductId { get; set; }
 
-        [Required]
-        [MaxLength(255)]
-        public required string Name { get; set; }
+    [Required]
+    [MaxLength(255)]
+    public required string Name { get; set; }
 
-        public required decimal PriceModifier { get; set; }
+    public required decimal PriceModifier { get; set; }
 
-        public required ProductState ProductState { get; set; }
+    public required ProductState ProductState { get; set; }
 
-        public required DateTime CreatedAt { get; set; }
+    public required DateTime CreatedAt { get; set; }
 
-        public required DateTime UpdatedAt { get; set; }
+    public required DateTime UpdatedAt { get; set; }
 
-    }
 }
