+++ Shopipy.ProductManagement/DTOs/ProductVariationRequestDTO.cs
@@ -1,11 +1,10 @@
 ﻿using Shopipy.Persistence.Models;
 
-namespace Shopipy.ProductManagement.DTOs
+namespace Shopipy.ProductManagement.DTOs;
+
+public class ProductVariationRequestDTO
 {
-    public class ProductVariationRequestDTO
-    {
-        public required string Name { get; set; }
-        public required decimal PriceModifier { get; set; }
-        public required ProductState ProductState { get; set; }
-    }
+    public required string Name { get; set; }
+    public required decimal PriceModifier { get; set; }
+    public required ProductState ProductState { get; set; }
 }
