+++ Shopipy.ProductManagement/DTOs/ProductVariationResponseDTO.cs
@@ -1,15 +1,14 @@
 ﻿using Shopipy.Persistence.Models;
 
-namespace Shopipy.ProductManagement.DTOs
+namespace Shopipy.ProductManagement.DTOs;
+
+public class ProductVariationResponseDTO
 {
-    public class ProductVariationResponseDTO
-    {
-        public int VariationId { get; set; }
-        public int ProductId { get; set; }
-        public string Name { get; set; } = string.Empty;
-        public decimal PriceModifier { get; set; }
-        public ProductState ProductState { get; set; }
-        public DateTime CreatedAt { get; set; }
-        public DateTime UpdatedAt { get; set; }
-    }
+    public int VariationId { get; set; }
+    public int ProductId { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public decimal PriceModifier { get; set; }
+    public ProductState ProductState { get; set; }
+    public DateTime CreatedAt { get; set; }
+    public DateTime UpdatedAt { get; set; }
 }
