+++ Shopipy.ProductManagement/DependencyInjection.cs
@@ -1,15 +1,14 @@
 ﻿using Shopipy.Shared.Services;
 using Shopipy.ProductManagement.Services;
 
-namespace Shopipy.ProductManagement
+namespace Shopipy.ProductManagement;
+
+public static class DependencyInjection
 {
-    public static class DependencyInjection
+    public static IServiceCollection AddProductManagement(this IServiceCollection products)
     {
-        public static IServiceCollection AddProductManagement(this IServiceCollection products)
-        {
-            products.AddScoped<IProductService, ProductService>();
-            products.AddScoped<IProductVariationService, ProductVariationService>();
-            return products;
-        }
+        products.AddScoped<IProductService, ProductService>();
+        products.AddScoped<IProductVariationService, ProductVariationService>();
+        return products;
     }
 }
