+++ Shopipy.ProductManagement/Mappings/ProductMappingProfile.cs
@@ -2,21 +2,20 @@
 using Shopipy.Persistence.Models;
 using Shopipy.ProductManagement.DTOs;
 
-namespace Shopipy.ProductManagement.Mappings
+namespace Shopipy.ProductManagement.Mappings;
+
+public class ProductMappingProfile : Profile
 {
-    public class ProductMappingProfile : Profile
+    public ProductMappingProfile()
     {
-        public ProductMappingProfile()
-        {
-            CreateMap<Product, ProductResponseDTO>();
+        CreateMap<Product, ProductResponseDTO>();
 
-            CreateMap<ProductRequestDTO, Product>()
-                .ForMember(dest => dest.BusinessId, opt => opt.Ignore())
-                .ForMember(dest => dest.ProductId, opt => opt.Ignore())
-                .ForMember(dest => dest.CreatedAt, opt => opt.Ignore())
-                .ForMember(dest => dest.UpdatedAt, opt => opt.Ignore());
-
-        }
+        CreateMap<ProductRequestDTO, Product>()
+            .ForMember(dest => dest.BusinessId, opt => opt.Ignore())
+            .ForMember(dest => dest.ProductId, opt => opt.Ignore())
+            .ForMember(dest => dest.CreatedAt, opt => opt.Ignore())
+            .ForMember(dest => dest.UpdatedAt, opt => opt.Ignore());
 
     }
+
 }
