+++ Shopipy.ProductManagement/Services/ProductService.cs
@@ -2,48 +2,47 @@
 using Shopipy.Persistence.Repositories;
 using Shopipy.Shared.Services;
 
-namespace Shopipy.ProductManagement.Services
+namespace Shopipy.ProductManagement.Services;
+
+public class ProductService(IGenericRepository<Product> _productRepository) : IProductService
 {
-    public class ProductService(IGenericRepository<Product> _productRepository) : IProductService
+    public async Task<Product> CreateProductAsync(Product product, int businessId)
     {
-        public async Task<Product> CreateProductAsync(Product product, int businessId)
-        {
-            product.BusinessId = businessId;
-            product.CreatedAt = DateTime.UtcNow;
-            product.UpdatedAt = DateTime.UtcNow;
+        product.BusinessId = businessId;
+        product.CreatedAt = DateTime.UtcNow;
+        product.UpdatedAt = DateTime.UtcNow;
 
-            var createdProduct = await _productRepository.AddAsync(product);
-            return createdProduct;
-        }
+        var createdProduct = await _productRepository.AddAsync(product);
+        return createdProduct;
+    }
 
-        public async Task<Product?> GetProductByIdAsync(int productId, int businessId)
-        {
-            var product = await _productRepository.GetByConditionAsync(p => p.ProductId == productId && p.BusinessId == businessId);
-            return product;
-        }
+    public async Task<Product?> GetProductByIdAsync(int productId, int businessId)
+    {
+        var product = await _productRepository.GetByConditionAsync(p => p.ProductId == productId && p.BusinessId == businessId);
+        return product;
+    }
 
-        public async Task<Product?> UpdateProductAsync(int productId, Product product, int businessId)
-        {
-            var existingProduct = await _productRepository.GetByConditionAsync(p => p.ProductId == productId && p.BusinessId == businessId);
-            if (existingProduct == null) return null;
+    public async Task<Product?> UpdateProductAsync(int productId, Product product, int businessId)
+    {
+        var existingProduct = await _productRepository.GetByConditionAsync(p => p.ProductId == productId && p.BusinessId == businessId);
+        if (existingProduct == null) return null;
 
-            existingProduct.Name = product.Name;
-            existingProduct.Description = product.Description;
-            existingProduct.BasePrice = product.BasePrice;
-            existingProduct.ProductState = product.ProductState;
-            existingProduct.CategoryId = product.CategoryId;
-            existingProduct.UpdatedAt = DateTime.UtcNow;
+        existingProduct.Name = product.Name;
+        existingProduct.Description = product.Description;
+        existingProduct.BasePrice = product.BasePrice;
+        existingProduct.ProductState = product.ProductState;
+        existingProduct.CategoryId = product.CategoryId;
+        existingProduct.UpdatedAt = DateTime.UtcNow;
 
-            var updatedProduct = await _productRepository.UpdateAsync(existingProduct);
-            return updatedProduct;
-        }
+        var updatedProduct = await _productRepository.UpdateAsync(existingProduct);
+        return updatedProduct;
+    }
 
-        public async Task<bool> DeleteProductAsync(int productId, int businessId)
-        {
-            var product = await _productRepository.GetByConditionAsync(p => p.ProductId == productId && p.BusinessId == businessId);
-            if (product == null) return false;
+    public async Task<bool> DeleteProductAsync(int productId, int businessId)
+    {
+        var product = await _productRepository.GetByConditionAsync(p => p.ProductId == productId && p.BusinessId == businessId);
+        if (product == null) return false;
 
-            return await _productRepository.DeleteAsync(product.ProductId);
-        }
+        return await _productRepository.DeleteAsync(product.ProductId);
     }
 }
